Refactoring: implementazione pattern observer per show dialog requests - Aggiunta interfaccia "DialogEventListener" - Aggiunta enum "DialogType" - Aggiunta classe "DialogEventManager" - Modifica di tutte le azioni per inviare show dialog requests - Modifica classe "RuleList" per inoltro di show dialog requests - Modifica classe "RuleManager" per gestione di show dialog requests - Rinominazione attributo e getter "eventManager" in "ruleEventManager" - Aggiunta unit test "DialogEventManagerTest" - Aggiunta classe helper per i test "FakeDialogEventListener" - Aggiunta test pattern observer di dialog in "RuleListTest" - Modifica FakeAction per supportare DialogEventManager Modifica "ExecuteExternalProgramAction" per non attendere fine processo

+++ src/softeng_project_group08/model/actions/MoveFileAction.java
@@ -6,9 +6,12 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import softeng_project_group08.model.DialogEventManager;
+import softeng_project_group08.model.DialogType;
 
 /**
  * Class that represents an action to move a file to a specified directory.
+ * Notifies listeners when there's a show dialog request.
  *
  * @author group08
  */
@@ -16,33 +19,51 @@ public class MoveFileAction implements Action {
 
     private File sourceFile;
     private File targetDirectory;
+    private DialogEventManager dialogEventManager;
 
     // Constructor to initialize the source file and target directory
     public MoveFileAction(File sourceFile, File targetDirectory) {
+        this.dialogEventManager = new DialogEventManager();
         this.sourceFile = sourceFile;
         this.targetDirectory = targetDirectory;
     }
 
     // Method to execute the file move action
     public void execute() {
+        // If the file or folder doesn't exist, notify the user with a dialog
+        if (!sourceFile.exists()) {
+            dialogEventManager.requestDialog(DialogType.ERROR,
+                    "MoveFileAction Error",
+                    "File to copy does not exist:\n" + sourceFile.getAbsolutePath());
+            return;
+        }
+        if (!targetDirectory.exists()) {
+            dialogEventManager.requestDialog(DialogType.ERROR,
+                    "MoveFileAction Error",
+                    "Target directory does not exist:\n" + targetDirectory.getAbsolutePath());
+            return;
+        }
         // Construct the path for the destination file within the target directory
         Path targetPath = targetDirectory.toPath().resolve(sourceFile.getName());
-
         try {
-            // Attempt to move the source file to the target path
             Files.move(sourceFile.toPath(), targetPath, StandardCopyOption.REPLACE_EXISTING);
-            // Print success message if the file was moved successfully
-            System.out.println("File moved successfully from " + sourceFile.getAbsolutePath()
-                    + " to " + targetPath);
+            System.out.println("The file has been copied successfully.");
         } catch (IOException ex) {
-            // If an IOException occurs during the move operation, print an error message and the stack trace
-            System.err.println("Error while moving the file: " + ex.getMessage());
-            ex.printStackTrace();
+            dialogEventManager.requestDialog(DialogType.ERROR,
+                    "MoveFileAction Error",
+                    "Generic file move error:\n" + sourceFile.getAbsolutePath()
+                    + "\n" + targetDirectory.getAbsolutePath());
         }
+    }
+
+    @Override
+    public DialogEventManager getDialogEventManager() {
+        return dialogEventManager;
     }
 
     @Override
     public String toString() {
         return "MoveFileAction:\n" + "sourceFile=\n" + sourceFile + "\ntargetDirectory=\n" + targetDirectory;
     }
+
 }
